aws-firestore-rds-pipeline/requirements.txt
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-google
google-cloud-firestore

